+++ src/json/store.rs
@@ -127,10 +127,11 @@ pub fn to_vec_pretty(font: &Font) -> crate::Result<Vec<u8>> {
 ///     Ok(())
 /// }
 /// ```
-pub fn to_writer<W: io::Write>(mut writer: W, font: &Font) -> io::Result<()> {
-    let json = serde_json::ser::to_string(&font)
-        .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
-    write!(writer, "{}", json)
+pub fn to_writer<W: io::Write>(mut writer: W, font: &Font) -> crate::Result<()> {
+    let json = serde_json::ser::to_string(&font).map_err(|e| {
+        crate::Error::UnsupportedValueEncoding { path: "json".to_owned(), value: e.to_string() }
+    })?;
+    write!(writer, "{}", json).map_err(Into::into)
 }
 
 /// Write JSON pretty format font.
@@ -156,8 +157,9 @@ pub fn to_writer<W: io::Write>(mut writer: W, font: &Font) -> io::Result<()> {
 ///     Ok(())
 /// }
 /// ```
-pub fn to_writer_pretty<W: io::Write>(mut writer: W, font: &Font) -> io::Result<()> {
-    let json = serde_json::ser::to_string_pretty(&font)
-        .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
-    write!(writer, "{}", json)
+pub fn to_writer_pretty<W: io::Write>(mut writer: W, font: &Font) -> crate::Result<()> {
+    let json = serde_json::ser::to_string_pretty(&font).map_err(|e| {
+        crate::Error::UnsupportedValueEncoding { path: "json".to_owned(), value: e.to_string() }
+    })?;
+    write!(writer, "{}", json).map_err(Into::into)
 }